requirements.txt
blinker==1.7.0
click==8.1.7
flask==3.0.2
Flask-Login==0.6.3
flask-marshmallow==1.2.0
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.20.2
marshmallow-sqlalchemy==1.0.0
packaging==23.2
SQLAlchemy==2.0.25
typing-extensions==4.9.0
werkzeug==3.0.1
zipp==3.17.0
psycopg2-binary==2.9.9
python-dotenv==0.19.1